requirements.txt
git+https://github.com/algorand/pyteal
python-dotenv